+++ packages/orq-rc/src/orq_ai_sdk/models/fileuploadop.py
@@ -93,5 +93,5 @@ class FileUploadResponseBody(BaseModel):
     workspace_id: str
     r"""The id of the resource"""
 
-    created: Optional[datetime] = parse_datetime("2025-07-28T12:49:31.256Z")
+    created: Optional[datetime] = parse_datetime("2025-07-29T08:09:24.692Z")
     r"""The date and time the resource was created"""