+++ packages/orq-rc/RELEASES.md
@@ -2129,3 +2129,13 @@ Based on:
 - [python v3.10.0-rc.23] packages/orq-rc
 ### Releases
 - [PyPI v3.10.0-rc.23] https://pypi.org/project/orq-ai-sdk/3.10.0-rc.23 - packages/orq-rc
+
+## 2025-07-29 08:10:51
+### Changes
+Based on:
+- OpenAPI Doc  
+- Speakeasy CLI 1.592.1 (2.667.0) https://github.com/speakeasy-api/speakeasy
+### Generated
+- [python v3.10.0-rc.24] packages/orq-rc
+### Releases
+- [PyPI v3.10.0-rc.24] https://pypi.org/project/orq-ai-sdk/3.10.0-rc.24 - packages/orq-rc